+++ rustls/src/error.rs
@@ -107,7 +107,7 @@ pub enum Error {
     NoApplicationProtocol,
 
     /// The MTU value supplied to the client was too small.
-    MtuError,
+    MtuTooSmall,
 }
 
 fn join<T: fmt::Debug>(items: &[T]) -> String {
@@ -157,7 +157,7 @@ impl fmt::Display for Error {
             Error::InvalidSct(ref err) => write!(f, "invalid certificate timestamp: {:?}", err),
             Error::FailedToGetCurrentTime => write!(f, "failed to get current time"),
             Error::FailedToGetRandomBytes => write!(f, "failed to get random bytes"),
-            Error::MtuError => write!(f, "the supplied MTU was too small"),
+            Error::MtuTooSmall => write!(f, "the supplied MTU was too small"),
             Error::General(ref err) => write!(f, "unexpected error: {}", err), // (please file a bug)
         }
     }
@@ -215,7 +215,7 @@ mod tests {
             Error::HandshakeNotComplete,
             Error::PeerSentOversizedRecord,
             Error::NoApplicationProtocol,
-            Error::MtuError,
+            Error::MtuTooSmall,
         ];
 
         for err in all {
